refactor(models): give the dialog a component folder and fns a kind folder (LFE-14948)
The dialog gained a test file, so it becomes a folder named after the
component. utils.ts and formValues.ts were functions outside a kind folder
(structure RFC rules 3 and 5); they move to fns/.
