+++ web/src/features/models/components/CloneModelButton.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@/src/components/ui/button";
 import { useHasProjectAccess } from "@/src/features/rbac/utils/checkProjectAccess";
-import { UpsertModelFormDialog } from "@/src/features/models/components/UpsertModelFormDialog";
+import { UpsertModelFormDialog } from "@/src/features/models/components/UpsertModelFormDialog/UpsertModelFormDialog";
 import { type GetModelResult } from "@/src/features/models/validation";
 
 export const CloneModelButton = ({
