+++ web/src/features/models/components/test-match/MatchedTierCard.tsx
@@ -8,7 +8,7 @@ import { Badge } from "@/src/components/ui/badge";
 import { useMemo } from "react";
 import { usePriceUnitMultiplier } from "@/src/features/models/hooks/usePriceUnitMultiplier";
 import Decimal from "decimal.js";
-import { getMaxDecimals } from "@/src/features/models/utils";
+import { getMaxDecimals } from "@/src/features/models/fns/getMaxDecimals";
 
 type MatchedTierCardProps = {
   tier: {
