+++ web/src/features/traces/components/ObservationDetailView/components/ModelBadge.tsx
@@ -6,7 +6,7 @@
 import { Badge } from "@/src/components/ui/badge";
 import { ExternalLinkIcon, PlusCircle } from "lucide-react";
 import Link from "next/link";
-import { UpsertModelFormDialog } from "@/src/features/models/components/UpsertModelFormDialog";
+import { UpsertModelFormDialog } from "@/src/features/models/components/UpsertModelFormDialog/UpsertModelFormDialog";
 
 export function ModelBadge({
   model,
